Renamed test init file
